Construct igraph network graph

Use this function to create an igraph graph object based on the topology .sif file given. It automatically sets various visualization graph properties and checks if the node names from the topology file are the same as in the models inside the given models.dir (if not NULL).

construct_network(topology.file, models.dir = NULL)

Arguments

topology.file
string. The name of the .sif file (can be a full path name).

models.dir
string. A dir with .gitsbe files/models. Default value: NULL. If specified, it is used for the validation of the node names.

Value

an igraph graph object representing the network as defined in the topology file

See also

Examples

topology.file = system.file("extdata", "example.sif", package = "emba", mustWork = TRUE)
net = construct_network(topology.file)