Get average link operator difference matrix based on the number of true positives

This function uses the get_avg_activity_diff_mat_based_on_tp_predictions function with the parameter models.link.operator as input in the place of models.stable.state, since the two matrices representing the two inputs have the same data format (rows represent models, columns represent nodes, and each value is a number in the [0,1] interval).

get_avg_link_operator_diff_mat_based_on_tp_predictions(
  models.synergies.tp,
  models.link.operator,
  penalty = 0
)

Arguments

models.synergies.tp
an integer vector of TP values. The names attribute must hold the models' names. Consider using the function calculate_models_synergies_tp.

models.link.operator
a data.frame (nxm) with n models and m nodes. The row names specify the models' names (same order as in the models.synergies.tp parameter) whereas the column names specify the network nodes (gene, proteins, etc.). Possible values for each model-node element are either 0 (AND NOT link operator), 1 (OR NOT link operator) or 0.5 if the node is not targeted by both activating and inhibiting regulators (no link operator).

penalty
value between 0 and 1 (inclusive). A value of 0 means no penalty and a value of 1 is the strickest possible penalty. Default value is 0. This penalty is used as part of a weighted term to the difference in a value of interest (e.g. activity or link operator difference) between two group of models, to account for the difference in the number of models from each respective model group.

Value

a matrix whose rows are vectors of average node link operator differences between two groups of models based on some kind of classification (e.g. number of TP predictions) and whose names are set in the rownames attribute of the matrix (usually denoting the different classification groups, e.g. (1,2) means the models that predicted 1 TP synergy vs the models that predicted 2 TP synergies, if the classification is done by number of TP predictions). The columns represent the network's node names. Values are in the [-1,1] interval.

Details

So, if a node has a value close to -1 it means that on average, this node's boolean equation has the AND NOT link operator in the 'good' models compared to the 'bad' ones while a value closer to 1 means that the node's boolean equation has mostly the OR NOT link operator in the 'good' models. A value closer to 0 indicates that the link operator in the node's boolean equation is not so much different between the 'good' and 'bad' models and so it won't not be a node of interest when searching for indicators of better performance (higher number of true positives) in the parameterization of the good models (the boolean equations). A value exactly equal to 0 can also mean that this node didn't not have a link operator in its boolean equation, again making it a non-important indicator of difference in model performance.

See also