Load the models boolean equation link operator data

Use this function to merge the link operator data used in the boolean equations of the models into a single data.frame object. Every boolean model is defined by a series of boolean equations in the form \(Target *= (Activator or Activator or...) and not (Inhibitor or Inhibitor or...)"\). The link operator can be either and not, or not or non-existent if the target has only activating regulators or only inhibiting ones (the not remains in the latter case). The models are loaded from .gitsbe files (and only these) that can be found inside the given models.dir directory.

get_link_operators_from_models_dir(
  models.dir,
  remove.equations.without.link.operator = TRUE
)

Arguments

models.dir
string. A directory path with .gitsbe files/models. Do not include the ending path character in the string (/).

remove.equations.without.link.operator
logical. Should we keep the nodes (columns in the returned matrix) which do not have both type of regulators (so no link operator)? Default value: TRUE (remove these nodes).

Value

a data.frame (nxm) with n models and m nodes. The row names specify the models' names whereas the column names specify the network nodes (gene, proteins, etc.). Possible values for each model-node element are either 0 (and not link operator), 1 (or not link operator) or 0.5 if the node is not targeted by both activating and inhibiting regulators (no link operator).

Examples

models.dir = system.file("extdata", "models", package = "emba", mustWork = TRUE)
models.link.operator = get_link_operators_from_models_dir(models.dir)
models.link.operator.with.extra.nodes =
  get_link_operators_from_models_dir(models.dir, FALSE)