Get observed synergies per cell line

Use this function to get the observed synergies from the respective files inside the given list of cell line directories.

get_observed_synergies_per_cell_line(cell.line.dirs, drug.combos)

Arguments

cell.line.dirs
a character vector of the cell line directories, in the form of {path}/cell_line_name. The cell line name directory should be different for each element of the vector as we use it to fill in the rownames of the result data.frame object. Inside each cell line directory we read the observed synergies from a file called observed_synergies (if it exists and is non-empty). This file has the names of the observed drug combinations, one in each line.

drug.combos
a character vector with elements the names of all the drug combinations that were tested in the analysis.

Value

a data.frame, whose columns represent the drug combinations tested and the rows the cell lines. Possible values for each cell line-drug combination element are either 1 (an observed synergy) or 0 (non-observed synergy).