Get synergy biomarkers from dir

This function reads the synergy biomarker files inside the given directory and merges the results into a data.frame which it returns. This functions should be used when the synergy biomarker results are in separate files inside the directory given (see biomarkers.dir parameter).

get_synergy_biomarkers_from_dir(
  predicted.synergies,
  biomarkers.dir,
  models.dir = NULL,
  node.names = NULL
)

Arguments

predicted.synergies
a character vector of the synergies (drug combination names) that were predicted by at least one of the models in the dataset.

biomarkers.dir
string. It specifies the full path name of the directory which holds the biomarker files (without the ending character /). The biomarker files must be formatted as: %drug.comb%_biomarkers_active or %drug.comb%_biomarkers_inhibited, where %drug.comb% is an element of the predicted.synergies vector.

models.dir
string. A directory with .gitsbe files/models. It's needed in order to call get_node_names.

node.names
a character vector which has the names of the nodes. If it's not NULL, then it will be used instead of the models.dir parameter. The node.names should include all the nodes that are reported as biomarkers in the biomarker files inside the biomarkers.dir directory. Note that the biomarker nodes in the files will be included in the returned data.frame object no matter the node.names specified. Default value: NULL.

Value

a data.frame, whose columns represent the network nodes and the rows the predicted synergies. Possible values for each synergy-node element are either 1 (active state biomarker), -1 (inhibited state biomarker) or 0 (not a biomarker or the node is not at all present in the network or the drug combination is not a synergistic one).