Get the refined x-axis values

This function returns the x-axis values that are going to be used by make_barplot_on_models_stats to render the bar plot.

get_x_axis_values(models.stats, there.is.one.NaN.category, cont.values)

Arguments

models.stats
table object, the result of using table on a (numeric) vector. Usually it represents some models statistics summary - counts for each TP prediction value for example.

there.is.one.NaN.category
logical. Is there one NaN category? (check is done before on the names attribute of the models.stats)

cont.values
logical. If TRUE, the values of the x-axis will be trimmed to 3 digits after the decimal point. Otherwise, they will be returned as they are.