Plot the graphs from an average state differences matrix

This function presents a convenient way to use the function plot_avg_state_diff_graph multiple times.

plot_avg_state_diff_graphs(net, diff.mat, layout = NULL)

Arguments

net
igraph graph object

diff.mat
a matrix whose rows are vectors of average node activity state differences between two groups of models based on some kind of classification (e.g. number of TP predictions) and whose names are set in the rownames attribute of the matrix (usually denoting the different classification groups, e.g. (1,2) means the models that predicted 1 TP synergy vs the models that predicted 2 TP synergies, if the classification is done by number of TP predictions). The columns represent the network's node names. Could be the result of using the function get_avg_activity_diff_mat_based_on_tp_predictions.

layout
a (nx2) numeric matrix of x-y coordinates (2 columns) for each of the nodes (n) in the net igraph object. If NULL, we use the default layout provided by layout_nicely.

See also