Update biomarker files for a specific synergy

This function gets the (previously-found or 'old') synergy biomarkers from their respective files and if any of these files are empty (no 'old' biomarkers found) or non-existent, the 'new' biomarkers (given as input vector parameters) are automatically saved. When the 'new' biomarkers share common nodes with the 'old' biomarkers, there exist 3 possible ways to combine the results, given by the method parameter. If no common nodes exist, no matter the method selected, the 'new' biomarkers are added to the 'old' ones.

update_biomarker_files(
  biomarkers.dir,
  drug.comb,
  biomarkers.active.new,
  biomarkers.inhibited.new,
  method = "replace"
)

Arguments

biomarkers.dir
string. It specifies the full path name of the directory (without the ending character /) which holds the biomarker files for the synergistic drug combination specified in the parameter drug.comb. The biomarker files must be formatted as: %drug.comb%_biomarkers_active or %drug.comb%_biomarkers_inhibited, where %drug.comb% is the value of the drug.comb parameter.

drug.comb
string. The drug combination (e.g. "A-B") that will be used to identify the related biomarker files.

biomarkers.active.new
a numeric vector whose names attribute includes the node names of the (newly found) active biomarkers for the specified synergy. The values of the vector are the average activity difference of each node, derived from a comparison between 2 different groups of models.

biomarkers.inhibited.new
a numeric vector whose names attribute includes the node names of the (newly found) inhibited biomarkers for the specified synergy. The values of the vector are the average activity difference of each node, derived from a comparison between 2 different groups of models.

method
string. It specifies the method to use to update the biomarker files when there are common nodes between the 'old' and 'new' biomarkers:

replace(DEFAULT): we discard the 'old' biomarkers and keep only the 'new' ones

prune.to.common: we keep only the common biomarkers

extend: we add to the 'old' set of biomarkers the extra ones from the 'new' set that are not non-common to the 'old' ones, extending thus the 'old' biomarker set